refactor(alerting): every "which silences" question in one module
Five reads select from `alert_silences`, and they were spread over three
modules: the as-code list in the silences repository, three screen reads
under `triage/`, and delivery's own in `suppression.ts`. They differ
mostly in their window test, which is the one part of a silence read worth
comparing, and no two of them could be compared without opening two files.
That is how two of them came to share the name `loadActiveSilences` while
meaning different things.

They now sit together in the silences repository, under a note that lists
each window test on one line. Reading them side by side turns up one
disagreement: the as-code list treats a window as half-open at both ends,
so a silence that closed exactly as the window opened did not cover it,
and the two screen reads treat it as closed. Nothing turns on that yet and
nothing here changes it; it is written down where the next reader will
meet both rules at once.

`triage/silences.ts` keeps what it was actually about, which is how a
silence reads on screen: the one in force for a rule, the row, the record.
`suppression.ts` keeps the decision and the defer. The row type follows
the queries that return it.

Reads and writes together matches the repository beside it: `rules` and
`delivery` each keep their list next to their create.
