+++ packages/app/src/data/alerting/triage/silences.ts
@@ -1,20 +1,19 @@
 /**
- * Silences as the triage screen reads them: which one is in force for a rule,
- * what it looks like on a row, and the record the detail lists.
+ * Silences as the screens read them: which one is in force for a rule, what
+ * it looks like on a row, and the record the two lists print.
+ *
+ * Which silences to fetch is a separate question, and every version of it
+ * lives together in the silences repository.
  */
-import { and, desc, eq, gt, gte, lte, or, sql } from "drizzle-orm";
 import {
   ruleSubject,
   silenceIsInForce,
   silenceSelects,
 } from "@/data/alerting/silences/matching";
+import type { SilenceRow } from "@/data/alerting/silences/repository";
 import type { AlertingMatcher } from "@/data/alerting/types";
-import { db } from "@/db/client";
-import { alertSilences } from "@/db/schema";
 import { formatElapsed, silenceImpact } from "./format";
 import type { AlertSilenceRecord, AlertSilenceView } from "./view";
-
-export type SilenceRow = typeof alertSilences.$inferSelect;
 
 /** What a silence did to delivery: notifications it is still sitting on, and
  *  ones it dropped for good. */
@@ -22,94 +21,6 @@ export type SilenceImpactCounts = { held: number; dropped: number };
 
 /** What a silence did to delivery when history has no row for it. */
 const NO_IMPACT: SilenceImpactCounts = { held: 0, dropped: 0 };
-
-/**
- * Every silence that has not closed yet, the ones still to start included: the
- * screens list a scheduled window as well as a muting one. A cancelled silence
- * has its window collapsed by `expireSilence`, so `ends_at > now()` already
- * excludes it.
- *
- * Not `loadActiveSilences`: delivery has a loader by that name and it means
- * the narrower thing, the silences in force this instant.
- */
-export async function loadOpenSilences(
-  organizationId: string,
-): Promise<SilenceRow[]> {
-  return db
-    .select()
-    .from(alertSilences)
-    .where(
-      and(
-        eq(alertSilences.organizationId, organizationId),
-        gt(alertSilences.endsAt, sql`now()`),
-      ),
-    );
-}
-
-/**
- * Silences for one rule whose window overlaps `[from, to]`, newest first.
- * Bounded by the window rather than by "active now" on purpose: the question
- * a silence list answers is usually "why did nobody hear about this", and by
- * then the silence responsible has often already expired.
- */
-export async function loadSilencesInWindow(
-  organizationId: string,
-  ruleId: string,
-  severity: string,
-  from: Date,
-  to: Date,
-): Promise<SilenceRow[]> {
-  const rows = await db
-    .select()
-    .from(alertSilences)
-    .where(
-      and(
-        eq(alertSilences.organizationId, organizationId),
-        lte(alertSilences.startsAt, to),
-        gte(alertSilences.endsAt, from),
-      ),
-    )
-    .orderBy(desc(alertSilences.startsAt));
-  const subject = ruleSubject(ruleId, severity);
-  return rows.filter((row) => silenceSelects(row.matchers, subject));
-}
-
-/**
- * What the Silences page lists: every silence still open, whatever the picked
- * range, plus the closed ones whose window overlaps it. The open ones are the
- * control surface and must not vanish because the reader is looking at last
- * week; the closed ones are evidence, and the range is what bounds evidence
- * on every other screen here.
- *
- * Newest window first. The page regroups into what is open and what has
- * closed, so this order only settles ties within a group.
- */
-export async function loadSilencesForPage(
-  organizationId: string,
-  from: Date,
-  to: Date,
-): Promise<SilenceRow[]> {
-  return db
-    .select()
-    .from(alertSilences)
-    .where(
-      and(
-        eq(alertSilences.organizationId, organizationId),
-        or(
-          gt(alertSilences.endsAt, sql`now()`),
-          and(lte(alertSilences.startsAt, to), gte(alertSilences.endsAt, from)),
-        ),
-      ),
-    )
-    .orderBy(desc(alertSilences.startsAt))
-    .limit(PAGE_LIMIT);
-}
-
-/** As many as `loadSilenceImpact` counts for in one read. Retention keeps
- *  closed silences for 90 days, and an Organization that writes more than
- *  this many in a range that wide has a different problem than a list that
- *  stops. */
-const PAGE_LIMIT = 200;
 
 function isWholeRuleSilence(matchers: AlertingMatcher[]): boolean {
   return matchers.every((m) => m.label === "rule");
